Move Collect back to ease review
