+++ packages/transformers/js/core/src/lib.rs
@@ -12,7 +12,6 @@ extern crate serde;
 extern crate serde_bytes;
 extern crate sha1;
 
-mod collect;
 mod decl_collector;
 mod dependency_collector;
 mod env_replacer;
@@ -45,15 +44,16 @@ use swc_ecmascript::transforms::{
 };
 use swc_ecmascript::visit::{FoldWith, VisitWith};
 
-use collect::{Collect, CollectResult};
 use decl_collector::*;
 use dependency_collector::*;
 use env_replacer::*;
 use fs::inline_fs;
 use global_replacer::GlobalReplacer;
-use hoist::{hoist, HoistResult};
+use hoist::{hoist, CollectResult, HoistResult};
 use modules::esm2cjs;
 use utils::{CodeHighlight, Diagnostic, DiagnosticSeverity, SourceLocation, SourceType};
+
+use crate::hoist::Collect;
 
 type SourceMapBuffer = Vec<(swc_common::BytePos, swc_common::LineCol)>;
